pyproject.toml
[project]
name = "img-compress"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.6",
    "pillow>=12.0.0",
]